requirements.txt
numpy
pandas
certifi
python-dotenv
pymongo[srv]==3.12
scikit-learn
dill
pyyaml
-e .
